+++ synphot/compat.py
@@ -19,6 +19,8 @@ else:
     HAS_DUST_EXTINCTION = True
 
 
-__all__ = ['ASTROPY_LT_4_0', 'HAS_SPECUTILS', 'HAS_DUST_EXTINCTION']
+__all__ = ['ASTROPY_LT_4_1', 'ASTROPY_LT_4_0', 'HAS_SPECUTILS',
+           'HAS_DUST_EXTINCTION']
 
+ASTROPY_LT_4_1 = not minversion(astropy, '4.1')
 ASTROPY_LT_4_0 = not minversion(astropy, '4.0')